+++ src/postulo/core/server_urls.py
@@ -17,6 +17,11 @@ urlpatterns = [
     path("email/test/", views.EmailTestView.as_view(), name="email_test"),
     path("plugins/", views.PluginsView.as_view(), name="plugins"),
     path("plugins/action/", views.PluginActionView.as_view(), name="plugin_action"),
+    path(
+        "plugins/repositories/",
+        views.PluginRepositoryView.as_view(),
+        name="plugin_repository",
+    ),
     path("logs/", views.LogsView.as_view(), name="logs"),
     path("capture/", views.CaptureView.as_view(), name="capture"),
     path("defaults/", views.DefaultsView.as_view(), name="defaults"),